lpc11xx::adc

Module cr

A/D Control Register. The ADCR register must be written to select the operating mode before A/D conversion can occur

Structs

BURST_W

Write proxy for field BURST

CLKDIV_W

Write proxy for field CLKDIV

CLKS_W

Write proxy for field CLKS

EDGE_W

Write proxy for field EDGE

SEL_W

Write proxy for field SEL

START_W

Write proxy for field START

Enums

BURST_A

Burst mode.

CLKS_A

This field selects the number of clocks used for each conversion in Burst mode, and the number of bits of accuracy of the result in the LS bits of ADDR, between 11 clocks (10 bits) and 4 clocks (3 bits).

EDGE_A

This bit is significant only when the START field contains 010-111. In these cases: Start conversion on a falling edge on the selected CAP/MAT signal.

START_A

When the BURST bit is 0, these bits control whether and when an A/D conversion is started.

Type Aliases

BURST_R

Reader of field BURST

CLKDIV_R

Reader of field CLKDIV

CLKS_R

Reader of field CLKS

EDGE_R

Reader of field EDGE

R

Reader of register CR

SEL_R

Reader of field SEL

START_R

Reader of field START

W

Writer for register CR