lpc11xx::adc::cr

Type Alias R

pub type R = R<u32, CR>;

Reader of register CR

Aliased Type

struct R { /* private fields */ }

Implementations

impl R

pub fn sel(&self) -> SEL_R

Bits 0:7 - Selects which of the AD7:0 pins is (are) to be sampled and converted. Bit 0 selects Pin AD0, bit 1 selects pin AD1,…, and bit 7 selects pin AD7. In software-controlled mode (BURST = 0), only one channel can be selected, i.e. only one of these bits should be 1. In hardware scan mode (BURST = 1), any numbers of channels can be selected, i.e any or all bits can be set to 1. If all bits are set to 0, channel 0 is selected automatically (SEL = 0x01).

pub fn clkdiv(&self) -> CLKDIV_R

Bits 8:15 - The APB clock (PCLK) is divided by CLKDIV +1 to produce the clock for the ADC, which should be less than or equal to 4.5 MHz. Typically, software should program the smallest value in this field that yields a clock of 4.5 MHz or slightly less, but in certain cases (such as a high-impedance analog source) a slower clock may be desirable.

pub fn burst(&self) -> BURST_R

Bit 16 - Burst mode.

pub fn clks(&self) -> CLKS_R

Bits 17:19 - This field selects the number of clocks used for each conversion in Burst mode, and the number of bits of accuracy of the result in the LS bits of ADDR, between 11 clocks (10 bits) and 4 clocks (3 bits).

pub fn start(&self) -> START_R

Bits 24:26 - When the BURST bit is 0, these bits control whether and when an A/D conversion is started.

pub fn edge(&self) -> EDGE_R

Bit 27 - This bit is significant only when the START field contains 010-111. In these cases: Start conversion on a falling edge on the selected CAP/MAT signal.