lpc11xx::adc::gdr

Type Alias W

pub type W = W<u32, GDR>;

Writer for register GDR

Aliased Type

struct W { /* private fields */ }

Implementations

impl W

pub fn v_vref(&mut self) -> V_VREF_W<'_>

Bits 6:15 - When DONE is 1, this field contains a binary fraction representing the voltage on the ADn pin selected by the SEL field, divided by the voltage on the VDD pin. Zero in the field indicates that the voltage on the ADn pin was less than, equal to, or close to that on VSS, while 0x3FF indicates that the voltage on ADn was close to, equal to, or greater than that on VREF.

pub fn chn(&mut self) -> CHN_W<'_>

Bits 24:26 - These bits contain the channel from which the result bits V_VREF were converted.

pub fn overrun(&mut self) -> OVERRUN_W<'_>

Bit 30 - This bit is 1 in burst mode if the results of one or more conversions was (were) lost and overwritten before the conversion that produced the result in the V_VREF bits.

pub fn done(&mut self) -> DONE_W<'_>

Bit 31 - This bit is set to 1 when an A/D conversion completes. It is cleared when this register is read and when the ADCR is written. If the ADCR is written while a conversion is still in progress, this bit is set and a new conversion is started.