lpc11xx::c_can

Module canif_mctrl

Message interface 1 message control

Structs

DLC_3_0_W

Write proxy for field DLC_3_0

EOB_W

Write proxy for field EOB

INTPND_W

Write proxy for field INTPND

MSGLST_W

Write proxy for field MSGLST

NEWDAT_W

Write proxy for field NEWDAT

RMTEN_W

Write proxy for field RMTEN

RXIE_W

Write proxy for field RXIE

TXIE_W

Write proxy for field TXIE

TXRQST_W

Write proxy for field TXRQST

UMASK_W

Write proxy for field UMASK

Enums

EOB_A

End of buffer.

INTPND_A

Interrupt pending.

MSGLST_A

Message lost (only valid for message objects in the direction receive).

NEWDAT_A

New data.

RMTEN_A

Remote enable.

RXIE_A

Receive interrupt enable.

TXIE_A

Transmit interrupt enable.

TXRQST_A

Transmit request.

UMASK_A

Use acceptance mask If UMASK is set to 1, the message object’s mask bits have to be programmed during initialization of the message object before MAGVAL is set to 1.

Type Aliases

DLC_3_0_R

Reader of field DLC_3_0

EOB_R

Reader of field EOB

INTPND_R

Reader of field INTPND

MSGLST_R

Reader of field MSGLST

NEWDAT_R

Reader of field NEWDAT

R

Reader of register CANIF%s_MCTRL

RMTEN_R

Reader of field RMTEN

RXIE_R

Reader of field RXIE

TXIE_R

Reader of field TXIE

TXRQST_R

Reader of field TXRQST

UMASK_R

Reader of field UMASK

W

Writer for register CANIF%s_MCTRL