lpc11xx::c_can::canif_mctrl

Type Alias W

pub type W = W<u32, CANIF_MCTRL>;

Writer for register CANIF%s_MCTRL

Aliased Type

struct W { /* private fields */ }

Implementations

impl W

pub fn dlc_3_0(&mut self) -> DLC_3_0_W<'_>

Bits 0:3 - Data length code The Data Length Code of a Message Object must be defined the same as in all the corresponding objects with the same identifier at other nodes. When the Message Handler stores a data frame, it will write the DLC to the value given by the received message. 0000 - 1000 = Data frame has 0 - 8 data bytes. 1001 - 1111 = Data frame has 8 data bytes.

pub fn eob(&mut self) -> EOB_W<'_>

Bit 7 - End of buffer.

pub fn txrqst(&mut self) -> TXRQST_W<'_>

Bit 8 - Transmit request.

pub fn rmten(&mut self) -> RMTEN_W<'_>

Bit 9 - Remote enable.

pub fn rxie(&mut self) -> RXIE_W<'_>

Bit 10 - Receive interrupt enable.

pub fn txie(&mut self) -> TXIE_W<'_>

Bit 11 - Transmit interrupt enable.

pub fn umask(&mut self) -> UMASK_W<'_>

Bit 12 - Use acceptance mask If UMASK is set to 1, the message object’s mask bits have to be programmed during initialization of the message object before MAGVAL is set to 1.

pub fn intpnd(&mut self) -> INTPND_W<'_>

Bit 13 - Interrupt pending.

pub fn msglst(&mut self) -> MSGLST_W<'_>

Bit 14 - Message lost (only valid for message objects in the direction receive).

pub fn newdat(&mut self) -> NEWDAT_W<'_>

Bit 15 - New data.