lpc11xx::c_can

Module canstat

Status register

Structs

BOFF_W

Write proxy for field BOFF

EPASS_W

Write proxy for field EPASS

EWARN_W

Write proxy for field EWARN

LEC_W

Write proxy for field LEC

RXOK_W

Write proxy for field RXOK

TXOK_W

Write proxy for field TXOK

Enums

BOFF_A

Busoff status.

EPASS_A

Error passive.

EWARN_A

Warning status.

LEC_A

Last error code Type of the last error to occur on the CAN bus.The LEC field holds a code which indicates the type of the last error to occur on the CAN bus. This field will be cleared to 0 when a message has been transferred (reception or transmission) without error. The unused code 111 may be written by the CPU to check for updates.

RXOK_A

Received a message successfully This bit is reset by the CPU. It is never reset by the CAN controller.

TXOK_A

Transmitted a message successfully This bit is reset by the CPU. It is never reset by the CAN controller.

Type Aliases

BOFF_R

Reader of field BOFF

EPASS_R

Reader of field EPASS

EWARN_R

Reader of field EWARN

LEC_R

Reader of field LEC

R

Reader of register CANSTAT

RXOK_R

Reader of field RXOK

TXOK_R

Reader of field TXOK

W

Writer for register CANSTAT