lpc11xx::c_can::canstat

Type Alias R

pub type R = R<u32, CANSTAT>;

Reader of register CANSTAT

Aliased Type

struct R { /* private fields */ }

Implementations

impl R

pub fn lec(&self) -> LEC_R

Bits 0:2 - Last error code Type of the last error to occur on the CAN bus.The LEC field holds a code which indicates the type of the last error to occur on the CAN bus. This field will be cleared to 0 when a message has been transferred (reception or transmission) without error. The unused code 111 may be written by the CPU to check for updates.

pub fn txok(&self) -> TXOK_R

Bit 3 - Transmitted a message successfully This bit is reset by the CPU. It is never reset by the CAN controller.

pub fn rxok(&self) -> RXOK_R

Bit 4 - Received a message successfully This bit is reset by the CPU. It is never reset by the CAN controller.

pub fn epass(&self) -> EPASS_R

Bit 5 - Error passive.

pub fn ewarn(&self) -> EWARN_R

Bit 6 - Warning status.

pub fn boff(&self) -> BOFF_R

Bit 7 - Busoff status.