lpc11xx::c_can::canstat

Type Alias W

pub type W = W<u32, CANSTAT>;

Writer for register CANSTAT

Aliased Type

struct W { /* private fields */ }

Implementations

impl W

pub fn lec(&mut self) -> LEC_W<'_>

Bits 0:2 - Last error code Type of the last error to occur on the CAN bus.The LEC field holds a code which indicates the type of the last error to occur on the CAN bus. This field will be cleared to 0 when a message has been transferred (reception or transmission) without error. The unused code 111 may be written by the CPU to check for updates.

pub fn txok(&mut self) -> TXOK_W<'_>

Bit 3 - Transmitted a message successfully This bit is reset by the CPU. It is never reset by the CAN controller.

pub fn rxok(&mut self) -> RXOK_W<'_>

Bit 4 - Received a message successfully This bit is reset by the CPU. It is never reset by the CAN controller.

pub fn epass(&mut self) -> EPASS_W<'_>

Bit 5 - Error passive.

pub fn ewarn(&mut self) -> EWARN_W<'_>

Bit 6 - Warning status.

pub fn boff(&mut self) -> BOFF_W<'_>

Bit 7 - Busoff status.