lpc11xx::ct16b0

Module ccr

Capture Control Register (CCR). The CCR controls which edges of the capture inputs are used to load the Capture Registers and whether or not an interrupt is generated when a capture takes place

Structs

CAP0FE_W

Write proxy for field CAP0FE

CAP0I_W

Write proxy for field CAP0I

CAP0RE_W

Write proxy for field CAP0RE

CAP1FE_W

Write proxy for field CAP1FE

CAP1I_W

Write proxy for field CAP1I

CAP1RE_W

Write proxy for field CAP1RE

Enums

CAP0FE_A

Capture on CT16Bn_CAP0 falling edge: a sequence of 1 then 0 on CT16Bn_CAP0 will cause CR0 to be loaded with the contents of TC.

CAP0I_A

Interrupt on CT16Bn_CAP0 event: a CR0 load due to a CT16Bn_CAP0 event will generate an interrupt.

CAP0RE_A

Capture on CT16Bn_CAP0 rising edge: a sequence of 0 then 1 on CT16Bn_CAP0 will cause CR0 to be loaded with the contents of TC.

CAP1FE_A

Capture on CT16Bn_CAP1 falling edge: a sequence of 1 then 0 on CT16Bn_CAP1 will cause CR1 to be loaded with the contents of TC.

CAP1I_A

Interrupt on CT16Bn_CAP1 event: a CR1 load due to a CT16Bn_CAP1 event will generate an interrupt.

CAP1RE_A

Capture on CT16Bn_CAP1 rising edge: a sequence of 0 then 1 on CT16Bn_CAP1 will cause CR1 to be loaded with the contents of TC.

Type Aliases

CAP0FE_R

Reader of field CAP0FE

CAP0I_R

Reader of field CAP0I

CAP0RE_R

Reader of field CAP0RE

CAP1FE_R

Reader of field CAP1FE

CAP1I_R

Reader of field CAP1I

CAP1RE_R

Reader of field CAP1RE

R

Reader of register CCR

W

Writer for register CCR