lpc11xx::ct16b0::ctcr

Type Alias W

pub type W = W<u32, CTCR>;

Writer for register CTCR

Aliased Type

struct W { /* private fields */ }

Implementations

impl W

pub fn ctm(&mut self) -> CTM_W<'_>

Bits 0:1 - Counter/Timer Mode. This field selects which rising PCLK edges can increment Timer’s Prescale Counter (PC), or clear PC and increment Timer Counter (TC).

pub fn cis(&mut self) -> CIS_W<'_>

Bits 2:3 - Count Input Select. In counter mode (when bits 1:0 in this register are not 00), these bits select which CAP pin is sampled for clocking. Note: If Counter mode is selected in the CTCR register, bits 2:0 in the Capture Control Register (CCR) must be programmed as 000.

pub fn encc(&mut self) -> ENCC_W<'_>

Bit 4 - Setting this bit to one enables clearing of the timer and the prescaler when the capture-edge event specified in bits 7:5 occurs.

pub fn selcc(&mut self) -> SELCC_W<'_>

Bits 5:7 - When bit 4 is one, these bits select which capture input edge will cause the timer and prescaler to be cleared. These bits have no effect when bit 4 is zero.