lpc11xx::ct16b0::emr

Type Alias R

pub type R = R<u32, EMR>;

Reader of register EMR

Aliased Type

struct R { /* private fields */ }

Implementations

impl R

pub fn em0(&self) -> EM0_R

Bit 0 - External Match 0. This bit reflects the state of output CT16B0_MAT0/CT16B1_MAT0, whether or not this output is connected to its pin. When a match occurs between the TC and MR0, this bit can either toggle, go LOW, go HIGH, or do nothing. Bits EMR[5:4] control the functionality of this output. This bit is driven to the CT16B0_MAT0/CT16B1_MAT0 pins if the match function is selected in the IOCON registers (0 = LOW, 1 = HIGH).

pub fn em1(&self) -> EM1_R

Bit 1 - External Match 1. This bit reflects the state of output CT16B0_MAT1/CT16B1_MAT1, whether or not this output is connected to its pin. When a match occurs between the TC and MR1, this bit can either toggle, go LOW, go HIGH, or do nothing. Bits EMR[7:6] control the functionality of this output. This bit is driven to the CT16B0_MAT1/CT16B1_MAT1 pins if the match function is selected in the IOCON registers (0 = LOW, 1 = HIGH).

pub fn em2(&self) -> EM2_R

Bit 2 - External Match 2. This bit reflects the state of output match channel 2, whether or not this output is connected to its pin. When a match occurs between the TC and MR2, this bit can either toggle, go LOW, go HIGH, or do nothing. Bits EMR[9:8] control the functionality of this output. Note that on counter/timer 0 this match channel is not pinned out. This bit is driven to the CT16B1_MAT2 pin if the match function is selected in the IOCON registers (0 = LOW, 1 = HIGH).

pub fn em3(&self) -> EM3_R

Bit 3 - External Match 3. This bit reflects the state of output of match channel 3. When a match occurs between the TC and MR3, this bit can either toggle, go LOW, go HIGH, or do nothing. Bits EMR[11:10] control the functionality of this output. There is no output pin available for this channel on either of the 16-bit timers.

pub fn emc0(&self) -> EMC0_R

Bits 4:5 - External Match Control 0. Determines the functionality of External Match 0.

pub fn emc1(&self) -> EMC1_R

Bits 6:7 - External Match Control 1. Determines the functionality of External Match 1.

pub fn emc2(&self) -> EMC2_R

Bits 8:9 - External Match Control 2. Determines the functionality of External Match 2.

pub fn emc3(&self) -> EMC3_R

Bits 10:11 - External Match Control 3. Determines the functionality of External Match 3.