lpc11xx::ct32b0

Module ctcr

Count Control Register (CTCR). The CTCR selects between Timer and Counter mode, and in Counter mode selects the signal and edge(s) for counting

Structs

CIS_W

Write proxy for field CIS

CTM_W

Write proxy for field CTM

ENCC_W

Write proxy for field ENCC

SELCC_W

Write proxy for field SELCC

Enums

CIS_A

Count Input Select. When bits 1:0 in this register are not 00, these bits select which CAP pin is sampled for clocking:.

CTM_A

Counter/Timer Mode. This field selects which rising PCLK edges can increment Timer’s Prescale Counter (PC), or clear PC and increment Timer Counter (TC). Timer Mode: every rising PCLK edge.

SELCC_A

When bit 4 is one, these bits select which capture input edge will cause the timer and prescaler to be cleared. These bits have no effect when bit 4 is zero.

Type Aliases

CIS_R

Reader of field CIS

CTM_R

Reader of field CTM

ENCC_R

Reader of field ENCC

R

Reader of register CTCR

SELCC_R

Reader of field SELCC

W

Writer for register CTCR