lpc11xx::ct32b0::ctcr

Type Alias R

pub type R = R<u32, CTCR>;

Reader of register CTCR

Aliased Type

struct R { /* private fields */ }

Implementations

impl R

pub fn ctm(&self) -> CTM_R

Bits 0:1 - Counter/Timer Mode. This field selects which rising PCLK edges can increment Timer’s Prescale Counter (PC), or clear PC and increment Timer Counter (TC). Timer Mode: every rising PCLK edge.

pub fn cis(&self) -> CIS_R

Bits 2:3 - Count Input Select. When bits 1:0 in this register are not 00, these bits select which CAP pin is sampled for clocking:.

pub fn encc(&self) -> ENCC_R

Bit 4 - Setting this bit to one enables clearing of the timer and the prescaler when the capture-edge event specified in bits 7:5 occurs.

pub fn selcc(&self) -> SELCC_R

Bits 5:7 - When bit 4 is one, these bits select which capture input edge will cause the timer and prescaler to be cleared. These bits have no effect when bit 4 is zero.