lpc11xx::gpio0::iev

Type Alias IEV8_R

pub type IEV8_R = R<bool, IEV8_A>;

Reader of field IEV8

Aliased Type

struct IEV8_R { /* private fields */ }

Implementations

impl IEV8_R

pub fn variant(&self) -> IEV8_A

Get enumerated values variant

pub fn is_falling(&self) -> bool

Checks if the value of the field is FALLING

pub fn is_rising(&self) -> bool

Checks if the value of the field is RISING