lpc11xx::i2c::stat

Type Alias STATUS_R

pub type STATUS_R = R<u8, STATUS_A>;

Reader of field Status

Aliased Type

struct STATUS_R { /* private fields */ }

Implementations

impl STATUS_R

pub fn variant(&self) -> Variant<u8, STATUS_A>

Get enumerated values variant

pub fn is_bus_error(&self) -> bool

Checks if the value of the field is BUS_ERROR

pub fn is_start_condition_transmitted(&self) -> bool

Checks if the value of the field is START_CONDITION_TRANSMITTED

pub fn is_repeated_start_condition_transmitted(&self) -> bool

Checks if the value of the field is REPEATED_START_CONDITION_TRANSMITTED

pub fn is_write_address_transmitted_ack(&self) -> bool

Checks if the value of the field is WRITE_ADDRESS_TRANSMITTED_ACK

pub fn is_write_address_transmitted_not_ack(&self) -> bool

Checks if the value of the field is WRITE_ADDRESS_TRANSMITTED_NOT_ACK

pub fn is_data_byte_transmitted_ack(&self) -> bool

Checks if the value of the field is DATA_BYTE_TRANSMITTED_ACK

pub fn is_data_byte_transmitted_not_ack(&self) -> bool

Checks if the value of the field is DATA_BYTE_TRANSMITTED_NOT_ACK

pub fn is_arbitration_lost(&self) -> bool

Checks if the value of the field is ARBITRATION_LOST

pub fn is_read_address_transmitted_ack(&self) -> bool

Checks if the value of the field is READ_ADDRESS_TRANSMITTED_ACK

pub fn is_read_address_transmitted_not_ack(&self) -> bool

Checks if the value of the field is READ_ADDRESS_TRANSMITTED_NOT_ACK

pub fn is_data_byte_received_ack(&self) -> bool

Checks if the value of the field is DATA_BYTE_RECEIVED_ACK

pub fn is_data_byte_received_not_ack(&self) -> bool

Checks if the value of the field is DATA_BYTE_RECEIVED_NOT_ACK

pub fn is_own_write_address_received_ack(&self) -> bool

Checks if the value of the field is OWN_WRITE_ADDRESS_RECEIVED_ACK

pub fn is_arbitration_lost_own_write_address_received_ack(&self) -> bool

Checks if the value of the field is ARBITRATION_LOST_OWN_WRITE_ADDRESS_RECEIVED_ACK

pub fn is_general_call_received_ack(&self) -> bool

Checks if the value of the field is GENERAL_CALL_RECEIVED_ACK

pub fn is_arbitration_lost_general_call_received_ack(&self) -> bool

Checks if the value of the field is ARBITRATION_LOST_GENERAL_CALL_RECEIVED_ACK

pub fn is_own_write_address_data_byte_received_ack(&self) -> bool

Checks if the value of the field is OWN_WRITE_ADDRESS_DATA_BYTE_RECEIVED_ACK

pub fn is_own_write_address_data_byte_received_not_ack(&self) -> bool

Checks if the value of the field is OWN_WRITE_ADDRESS_DATA_BYTE_RECEIVED_NOT_ACK

pub fn is_general_call_data_byte_received_ack(&self) -> bool

Checks if the value of the field is GENERAL_CALL_DATA_BYTE_RECEIVED_ACK

pub fn is_general_call_data_byte_received_not_ack(&self) -> bool

Checks if the value of the field is GENERAL_CALL_DATA_BYTE_RECEIVED_NOT_ACK

pub fn is_stop_or_repeated_start_condition_received_while_addressed( &self, ) -> bool

Checks if the value of the field is STOP_OR_REPEATED_START_CONDITION_RECEIVED_WHILE_ADDRESSED

pub fn is_own_read_address_ack(&self) -> bool

Checks if the value of the field is OWN_READ_ADDRESS_ACK

pub fn is_arbitration_lost_own_read_address_received_ack(&self) -> bool

Checks if the value of the field is ARBITRATION_LOST_OWN_READ_ADDRESS_RECEIVED_ACK

pub fn is_slave_data_byte_transmitted_ack(&self) -> bool

Checks if the value of the field is SLAVE_DATA_BYTE_TRANSMITTED_ACK

pub fn is_slave_data_byte_transmitted_not_ack(&self) -> bool

Checks if the value of the field is SLAVE_DATA_BYTE_TRANSMITTED_NOT_ACK

pub fn is_last_data_byte_transmitted_ack_received(&self) -> bool

Checks if the value of the field is LAST_DATA_BYTE_TRANSMITTED_ACK_RECEIVED

pub fn is_not_available(&self) -> bool

Checks if the value of the field is NOT_AVAILABLE