lpc11xx::iocon::iocon_pio0_1

Type Alias OPEN_DRAIN_R

pub type OPEN_DRAIN_R = R<bool, OPEN_DRAIN_A>;

Reader of field OPEN_DRAIN

Aliased Type

struct OPEN_DRAIN_R { /* private fields */ }

Implementations

impl OPEN_DRAIN_R

pub fn variant(&self) -> OPEN_DRAIN_A

Get enumerated values variant

pub fn is_gpio_output(&self) -> bool

Checks if the value of the field is GPIO_OUTPUT

pub fn is_open_drain_output(&self) -> bool

Checks if the value of the field is OPEN_DRAIN_OUTPUT