lpc11xx

Module spi0

Serial Peripheral Interface

Modules

cpsr

Clock Prescale Register

cr0

Control Register 0. Selects the serial clock rate, bus type, and data size

cr1

Control Register 1. Selects master/slave and other modes

dr

Data Register. Writes fill the transmit FIFO, and reads empty the receive FIFO

icr

SSPICR Interrupt Clear Register

imsc

Interrupt Mask Set and Clear Register

mis

Masked Interrupt Status Register

ris

Raw Interrupt Status Register

sr

Status Register

Structs

RegisterBlock

Register block

Type Aliases

CPSR

Clock Prescale Register

CR0

Control Register 0. Selects the serial clock rate, bus type, and data size

CR1

Control Register 1. Selects master/slave and other modes

DR

Data Register. Writes fill the transmit FIFO, and reads empty the receive FIFO

ICR

SSPICR Interrupt Clear Register

IMSC

Interrupt Mask Set and Clear Register

MIS

Masked Interrupt Status Register

RIS

Raw Interrupt Status Register

SR

Status Register