lpc11xx::spi0::cr0

Type Alias W

pub type W = W<u32, CR0>;

Writer for register CR0

Aliased Type

struct W { /* private fields */ }

Implementations

impl W

pub fn dss(&mut self) -> DSS_W<'_>

Bits 0:3 - Data Size Select. This field controls the number of bits transferred in each frame. Values 0000-0010 are not supported and should not be used.

pub fn frf(&mut self) -> FRF_W<'_>

Bits 4:5 - Frame Format.

pub fn cpol(&mut self) -> CPOL_W<'_>

Bit 6 - Clock Out Polarity. This bit is only used in SPI mode.

pub fn cpha(&mut self) -> CPHA_W<'_>

Bit 7 - Clock Out Phase. This bit is only used in SPI mode.

pub fn scr(&mut self) -> SCR_W<'_>

Bits 8:15 - Serial Clock Rate. The number of prescaler output clocks per bit on the bus, minus one. Given that CPSDVSR is the prescale divider, and the APB clock PCLK clocks the prescaler, the bit frequency is PCLK / (CPSDVSR X [SCR+1]).