lpc11xx::spi0::cr1

Type Alias W

pub type W = W<u32, CR1>;

Writer for register CR1

Aliased Type

struct W { /* private fields */ }

Implementations

impl W

pub fn lbm(&mut self) -> LBM_W<'_>

Bit 0 - Loop Back Mode.

pub fn sse(&mut self) -> SSE_W<'_>

Bit 1 - SPI Enable.

pub fn ms(&mut self) -> MS_W<'_>

Bit 2 - Master/Slave Mode.This bit can only be written when the SSE bit is 0.

pub fn sod(&mut self) -> SOD_W<'_>

Bit 3 - Slave Output Disable. This bit is relevant only in slave mode (MS = 1). If it is 1, this blocks this SPI controller from driving the transmit data line (MISO).