lpc11xx::syscon::wdtoscctrl

Type Alias FREQSEL_R

pub type FREQSEL_R = R<u8, FREQSEL_A>;

Reader of field FREQSEL

Aliased Type

struct FREQSEL_R { /* private fields */ }

Implementations

impl FREQSEL_R

pub fn variant(&self) -> Variant<u8, FREQSEL_A>

Get enumerated values variant

pub fn is_0_5_mhz(&self) -> bool

Checks if the value of the field is _0_5_MHZ

pub fn is_0_8_mhz(&self) -> bool

Checks if the value of the field is _0_8_MHZ

pub fn is_1_1_mhz(&self) -> bool

Checks if the value of the field is _1_1_MHZ

pub fn is_1_4_mhz(&self) -> bool

Checks if the value of the field is _1_4_MHZ

pub fn is_1_6_mhz(&self) -> bool

Checks if the value of the field is _1_6_MHZ

pub fn is_1_8_mhz(&self) -> bool

Checks if the value of the field is _1_8_MHZ

pub fn is_2_0_mhz(&self) -> bool

Checks if the value of the field is _2_0_MHZ

pub fn is_2_2_mhz(&self) -> bool

Checks if the value of the field is _2_2_MHZ

pub fn is_2_4_mhz(&self) -> bool

Checks if the value of the field is _2_4_MHZ

pub fn is_2_6_mhz(&self) -> bool

Checks if the value of the field is _2_6_MHZ

pub fn is_2_7_mhz(&self) -> bool

Checks if the value of the field is _2_7_MHZ

pub fn is_2_9_mhz(&self) -> bool

Checks if the value of the field is _2_9_MHZ

pub fn is_3_1_mhz(&self) -> bool

Checks if the value of the field is _3_1_MHZ

pub fn is_3_2_mhz(&self) -> bool

Checks if the value of the field is _3_2_MHZ

pub fn is_3_4_mhz(&self) -> bool

Checks if the value of the field is _3_4_MHZ