lpc11xx::uart::iir

Type Alias INTID_R

pub type INTID_R = R<u8, INTID_A>;

Reader of field INTID

Aliased Type

struct INTID_R { /* private fields */ }

Implementations

impl INTID_R

pub fn variant(&self) -> Variant<u8, INTID_A>

Get enumerated values variant

pub fn is_rls(&self) -> bool

Checks if the value of the field is RLS

pub fn is_rda(&self) -> bool

Checks if the value of the field is RDA

pub fn is_cti(&self) -> bool

Checks if the value of the field is CTI

pub fn is_thre(&self) -> bool

Checks if the value of the field is THRE

pub fn is_modem(&self) -> bool

Checks if the value of the field is MODEM