lpc11xx::uart::lsr

Type Alias BI_R

pub type BI_R = R<bool, BI_A>;

Reader of field BI

Aliased Type

struct BI_R { /* private fields */ }

Implementations

impl BI_R

pub fn variant(&self) -> BI_A

Get enumerated values variant

pub fn is_inactive(&self) -> bool

Checks if the value of the field is INACTIVE

pub fn is_active(&self) -> bool

Checks if the value of the field is ACTIVE