lpc11xx::uart::mcr

Type Alias W

pub type W = W<u32, MCR>;

Writer for register MCR

Aliased Type

struct W { /* private fields */ }

Implementations

impl W

pub fn dtrc(&mut self) -> DTRC_W<'_>

Bit 0 - DTR Control. Source for modem output pin, DTR. This bit reads as 0 when modem loopback mode is active.

pub fn rtsc(&mut self) -> RTSC_W<'_>

Bit 1 - RTS Control. Source for modem output pin RTS. This bit reads as 0 when modem loopback mode is active.

pub fn lms(&mut self) -> LMS_W<'_>

Bit 4 - Loopback Mode Select. The modem loopback mode provides a mechanism to perform diagnostic loopback testing. Serial data from the transmitter is connected internally to serial input of the receiver. Input pin, RXD, has no effect on loopback and output pin, TXD is held in marking state. The four modem inputs (CTS, DSR, RI and DCD) are disconnected externally. Externally, the modem outputs (RTS, DTR) are set inactive. Internally, the four modem outputs are connected to the four modem inputs. As a result of these connections, the upper four bits of the MSR will be driven by the lower four bits of the MCR rather than the four modem inputs in normal mode. This permits modem status interrupts to be generated in loopback mode by writing the lower four bits of MCR.

pub fn rtsen(&mut self) -> RTSEN_W<'_>

Bit 6 - RTS flow control.

pub fn ctsen(&mut self) -> CTSEN_W<'_>

Bit 7 - CTS flow control.