lpc11xx

Module wwdt

Windowed Watchdog Timer

Modules

wdfeed

Watchdog feed sequence register. Writing 0xAA followed by 0x55 to this register reloads the Watchdog timer with the value contained in WDTC

wdmod

Watchdog mode register. This register contains the basic mode and status of the Watchdog Timer

wdtc

Watchdog timer constant register. This register determines the time-out value

wdtv

Watchdog timer value register. This register reads out the current value of the Watchdog timer

wdwarnint

Watchdog Warning Interrupt compare value

wdwindow

Watchdog Window compare value

Structs

RegisterBlock

Register block

Type Aliases

WDFEED

Watchdog feed sequence register. Writing 0xAA followed by 0x55 to this register reloads the Watchdog timer with the value contained in WDTC

WDMOD

Watchdog mode register. This register contains the basic mode and status of the Watchdog Timer

WDTC

Watchdog timer constant register. This register determines the time-out value

WDTV

Watchdog timer value register. This register reads out the current value of the Watchdog timer

WDWARNINT

Watchdog Warning Interrupt compare value

WDWINDOW

Watchdog Window compare value