lpc11xx::wwdt

Module wdmod

Watchdog mode register. This register contains the basic mode and status of the Watchdog Timer

Structs

WDEN_W

Write proxy for field WDEN

WDINT_W

Write proxy for field WDINT

WDPROTECT_W

Write proxy for field WDPROTECT

WDRESET_W

Write proxy for field WDRESET

WDTOF_W

Write proxy for field WDTOF

Enums

WDEN_A

Watchdog enable bit. This bit is Set Only. Setting this bit to one also locks the watchdog clock source. Once the watchdog timer is enabled, the watchdog timer clock source cannot be changed. If the watchdog timer is needed in Deep-sleep mode, the watchdog clock source must be changed to the watchdog oscillator before setting this bit to one.

WDPROTECT_A

Watchdog update mode. This bit is Set Only.

WDRESET_A

Watchdog reset enable bit. This bit is Set Only.

Type Aliases

R

Reader of register WDMOD

W

Writer for register WDMOD

WDEN_R

Reader of field WDEN

WDINT_R

Reader of field WDINT

WDPROTECT_R

Reader of field WDPROTECT

WDRESET_R

Reader of field WDRESET

WDTOF_R

Reader of field WDTOF